requirements.txt
streamlit
scikit-learn
nltk
spacy
python-docx
pdfplumber
PyMuPDF
wordcloud